refactor: 날씨 관련 유틸리티 함수 통합 및 불필요한 파일 삭제 (#62)
